% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pystg_is_available}
\alias{pystg_is_available}
\title{Check whether STG Python package is available and can be loaded}
\usage{
pystg_is_available()
}
\value{
No return value, called for side effects
}
\description{
This is used to avoid running tests on CRAN
}
