% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_stg.R
\name{stg}
\alias{stg}
\title{STochastic Gates (STG) based COX Proportional Hazards Regression Model}
\usage{
stg(data, ...)
}
\arguments{
\item{data}{string
path to the input survival data}

\item{hidden_layers_node}{vector of integers,optional,default:c(60, 20, 3)
architecture vector of the neural network}
}
\description{
coxstg extends cox proportional hazards regression model by using
a non-linear neural network feature selection model
(STochastic Gates (STG)), as described in Yamada, et al, ICML 2020.
}
\examples{
dataset <- "/data/jyc/cox_data/dataset/gaussian/gaussian_survival_data.h5"
coxstg(dataset,hidden_layers_node = c(60, 30, 20, 3),num_epoch = 3000,output_dir = "~/results/")
}
